Authorization - Protecting Routes and Verifying Current User

+++ routers/api/post.py
@@ -5,6 +5,7 @@ from sqlalchemy import select
 from sqlalchemy.ext.asyncio import AsyncSession
 from sqlalchemy.orm import selectinload
 
+from authetication.auth import CurrentUser
 from databases import models
 from databases.database import get_db
 from schemas.posts_schema import postCreate, postResponse, postUpdate
@@ -19,23 +20,12 @@ router = APIRouter()
     status_code=status.HTTP_201_CREATED,
 )
 
-async def create_post(post: postCreate, db: Annotated[AsyncSession, Depends(get_db)]):
-    result = await db.execute(
-        select(models.User)
-        .options(selectinload(models.User.posts))
-        .where(models.User.id == post.user_id)
-    )
-    user = result.scalars().first()
-    if not user:
-        raise HTTPException(
-            status_code=status.HTTP_404_NOT_FOUND,
-            detail="User not found"
-        )
+async def create_post(post: postCreate, current_user: CurrentUser, db: Annotated[AsyncSession, Depends(get_db)]):
 
     new_post = models.Post(
         title=post.title,
         content=post.content,
-        user_id=post.user_id,
+        user_id=current_user.id,
     )
     db.add(new_post)
     await db.commit()
@@ -74,15 +64,15 @@ async def get_post(post_id: int , db: Annotated[AsyncSession, Depends(get_db)]):
 
 
 @router.put("/{post_id}", response_model=postResponse)
-async def update_post_full(post_id: int, post_data: postCreate, db: Annotated[AsyncSession, Depends(get_db)]):
+async def update_post_full(post_id: int, post_data: postCreate, current_user: CurrentUser, db: Annotated[AsyncSession, Depends(get_db)]):
     post = await get_post(post_id, db)
 
-    if post_data.user_id != post.user_id:
+    if post.user_id != current_user.id:
         raise HTTPException(
             status_code=status.HTTP_403_FORBIDDEN,
-            detail="You are not authorized to update this post",
+            details = "Not authorized to update this post"
         )
-    
+
     post.title = post_data.title
     post.content = post_data.content
     post.user_id = post_data.user_id
@@ -91,8 +81,15 @@ async def update_post_full(post_id: int, post_data: postCreate, db: Annotated[As
     return post
 
 @router.patch("/{post_id}", response_model=postResponse)
-async def update_post_partial(post_id: int, post_data: postUpdate, db: Annotated[AsyncSession, Depends(get_db)]):
+async def update_post_partial(post_id: int, post_data: postUpdate, current_user: CurrentUser, db: Annotated[AsyncSession, Depends(get_db)]):
     post = await get_post(post_id, db)
+
+    if post.user_id != current_user.id:
+            raise HTTPException(
+                status_code=status.HTTP_403_FORBIDDEN,
+                detail = "Not authorized to update this post"
+            )
+    
 
     update_data = post_data.model_dump(exclude_unset=True)
     for key, value in update_data.items():
@@ -103,8 +100,16 @@ async def update_post_partial(post_id: int, post_data: postUpdate, db: Annotated
 
 
 @router.delete("/{post_id}", status_code=status.HTTP_204_NO_CONTENT)
-async def delete_post(post_id: int, db: Annotated[AsyncSession, Depends(get_db)]):
+async def delete_post(post_id: int, current_user: CurrentUser, db: Annotated[AsyncSession, Depends(get_db)]):
     post = await get_post(post_id, db)
+
+    if post.user_id != current_user.id:
+        raise HTTPException(
+            status_code=status.HTTP_403_FORBIDDEN,
+            details = "Not authorized to delete this post"
+        )
+
+
     await db.delete(post)
     await db.commit()
